+++ freppledb/common/static/js/frepple.js
@@ -393,6 +393,10 @@ jQuery.extend($.fn.fmatter, {
 
   duration : formatDuration,
 
+  image : function(cellvalue, options, rowdata) {
+    return '<img style="border-radius: 50%; width:40px; height:40px" src="/uploads/' + cellvalue + '">';
+  },
+  
   admin : function(cellvalue, options, rowdata) {
     if (cellvalue === undefined || cellvalue === '' || cellvalue === null)
       return '';