+++ doc/release-notes.rst
@@ -9,6 +9,8 @@ Release notes
 - Updated the list of default fields in all reports. By default we keep the
   reports now as lean as possible. More advanced fields are hidden by default.
 
+- User can now upload an avatar image.
+
 6.9.0 (2020/11/07)
 ==================